Fix build with pointer-typed NULL
If we switch NULL to be pointer-typed (rather than a plain int zero, as
it is currently), this reveals some improper usage of NULL where it is
being assigned to a non-pointer lvalue. We can't switch the type of NULL
as this breaks some matches, but we can fix these improper uses.

+++ src/melee/ft/chara/ftCommon/ftCo_Attack100.c
@@ -1130,7 +1130,7 @@ void fn_800D8378(Fighter_GObj* gobj)
 {
     Fighter* fp = GET_FIGHTER(gobj);
 
-    if ((u32) fp->item_gobj != NULL) {
+    if (fp->item_gobj != NULL) {
         ftCommon_8007E7E4(gobj, 1);
     }
 }
